A lot is going on here, form top to bottom:
I imported some Ring middleware. These modify the request and response on the fly.
Note that wrap-reload and wrap-stacktrace are for debuging only.

I added the middleware to the Moustache app.

I added routes. The first route matches and WikiLink and binds it to
title. The second one redirects all other links to the MainPage.

The #' syntax is for getting the var instead of the value, to make
reloading work.

At this point, visiting the same url should redirect to /MainPage and
display "hello tester"

Changing this text does not require a server restart, so we can keep it
running from now on.