I lied, to add a new dependency, you need to restart the server.
Hiccup is a DSL for generating HTML.

I defined a HTML template and a function for showing it that takes a
request and a title.

Note that I used the underscore to denote we're not using the request.

Delegate is a HOF that returns a function with the title alreadu
supplied. Moustache supplies the request.

Try visiting /FooBar now.