I added Clutch as a dependency. Clutch is a library for CouchDB.
Install CouchDB and use Futon to create a wiki database and insert a
"MainPage" document with a "content" key.

I defined my own Ring middleware that takes a handler and returns
another function that calls the old handler in the context of our
database.

The show function now gets a document from the database and passes its
content to page.