You can now edit and create pages.
Page now takes a revision, which is used in the web form to update.

The update function does use the request object, and uses destructuring
to extracts form data as parsed by wrap-params.

Depending if a revision was supplied, a new document is created or an
existing one updated. Then, the page is shown.

Note how Moustache now delegates POST requests to update and GET
requests to show.